Modif du diagramme

+++ GestionStockArticle/src/com/controller/EconomicDeliveryModel.java
@@ -4,14 +4,14 @@ import java.lang.Math;
 
 public class EconomicDeliveryModel extends  DeliveryModel {
     private int totalUnitPerYear;
-    private int launchCommandPrice;
+    private double launchCommandPrice;
 
     /**
      * A class which defines a delivery model for an article
      * @param totalUnitPerYear the total unit of the product sold per year
      * @param launchCommandPrice  the the launch command price of one product of the Article
      */
-    public EconomicDeliveryModel(int totalUnitPerYear, int launchCommandPrice,  int stockMinimum) {
+    public EconomicDeliveryModel(int totalUnitPerYear, double launchCommandPrice,  int stockMinimum) {
         super(stockMinimum, 0);
         this.totalUnitPerYear = totalUnitPerYear;
         this.launchCommandPrice = launchCommandPrice;
